build: extract the shared tsdown configuration into one factory
The four output configurations are the same for every package apart from the
bundle name, the global variable the bundle defines, and which dependencies stay
outside the bundle. Move the common part into tsdown.config.base.js and pass the
rest in, mirroring how rollup.config.base.js already worked.

The per-package file is now JavaScript rather than TypeScript, which matches the
Rollup files it replaces and avoids Node.js reparsing the config on every build.

Also keeps every package import external while generating declarations. Without
that, declarations for types reaching us through a dependency were copied into
dist/types under a node_modules directory, which would republish another
project's types as if they were ours.
