--- packages/web-tracing/tsdown.config.ts
@@ -1,78 +1,0 @@
-import { defineConfig } from 'tsdown';
-
-// Spike configuration. The output layout below is a published contract:
-// - `main`, `module` and `types` in package.json point into dist/cjs, dist/esm and dist/types
-// - dist/bundle/faro-web-tracing.iife.js is served from unpkg and is documented in
-//   docs/sources/tutorials/use-cdn-library.md
-// Keep every path exactly as it is today.
-
-// Every source file is an entry, not just index.ts. Anything less lets Rolldown tree-shake exports
-// that are not reachable from index.ts, which would make the emitted declarations disagree with the
-// emitted JavaScript for anyone importing a module directly.
-const sources = ['./src/**/*.ts', '!./src/**/*.test.ts'];
-
-const shared = {
-  entry: sources,
-  platform: 'browser',
-  target: 'es2015',
-  sourcemap: true,
-  hash: false,
-} as const;
-
-export default defineConfig([
-  // CommonJS, one output file per source file, mirroring the tsc build.
-  {
-    ...shared,
-    format: 'cjs',
-    unbundle: true,
-    outDir: './dist/cjs',
-    outExtensions: () => ({ js: '.js' }),
-    dts: false,
-  },
-  // ES modules, one output file per source file.
-  {
-    ...shared,
-    format: 'esm',
-    unbundle: true,
-    outDir: './dist/esm',
-    outExtensions: () => ({ js: '.js' }),
-    dts: false,
-  },
-  // Declarations only, kept in their own directory to match the current layout.
-  // tsconfig.esm.json sets isolatedDeclarations, so these are generated by Oxc rather than by the
-  // TypeScript compiler.
-  {
-    ...shared,
-    format: 'esm',
-    unbundle: true,
-    outDir: './dist/types',
-    tsconfig: './tsconfig.esm.json',
-    outExtensions: () => ({ dts: '.d.ts' }),
-    // Declaration maps are not emitted, so do not leave a sourceMappingURL comment pointing at a
-    // file that does not exist.
-    sourcemap: false,
-    dts: { emitDtsOnly: true, sourcemap: false },
-  },
-  // Minified IIFE bundle for the content delivery network.
-  {
-    ...shared,
-    format: 'iife',
-    entry: { 'faro-web-tracing': './src/index.ts' },
-    outDir: './dist/bundle',
-    globalName: 'GrafanaFaroWebTracing',
-    // tsdown externalises everything listed in `dependencies` by default. For the bundle we want
-    // the opposite: only faro-web-sdk stays external (it is expected as a global), and the
-    // OpenTelemetry packages are inlined, which is what the Rollup build does today.
-    deps: {
-      neverBundle: ['@grafana/faro-web-sdk'],
-      alwaysBundle: [/^@opentelemetry\//],
-    },
-    outputOptions: {
-      globals: {
-        '@grafana/faro-web-sdk': 'GrafanaFaroWebSdk',
-      },
-    },
-    minify: true,
-    dts: false,
-  },
-]);
